mood detection implemented
I got some code for using a webcam to look at your face, and detect your mood., so I implemented that to the chatbot. If you ask it "do you see me" or "can you see me", it will run the code, and tell you what it thinks your mood is.